fix: call ResetMouseDragDelta
If mouse is clicked at 0,0 and dragged to x, y; we get x, y when we call
GetMouseDragDelta. After rotating for x, y; we need to call ResetMouseDragDelta to get
new values when mouse is dragged from x, y to a, b else we will get a, b
rather than a-x,b-y.